+++ formula/src/main/java/com/instacart/formula/internal/DeferredTransition.kt
@@ -1,0 +1,21 @@
+package com.instacart.formula.internal
+
+import com.instacart.formula.Transition
+
+/**
+ * A deferred transition contains an [event] and a related [transition] which can
+ * be executed using the [execute] function. If the formula is ready, it will be
+ * executed immediately. Otherwise, it will be added to the queue and executed
+ * when formula is ready. If the formula is re-evaluated and [listener] is disabled
+ * before the transition is executed, this transition will be ignored.
+ */
+class DeferredTransition<Input, State, EventT> internal constructor(
+    private val listener: ListenerImpl<Input, State, EventT>,
+    private val transition: Transition<Input, State, EventT>,
+    private val event: EventT,
+) {
+
+    fun execute() {
+        listener.snapshotImpl?.dispatch(transition, event)
+    }
+}