+++ formula/src/main/java/com/instacart/formula/internal/ListenerImpl.kt
@@ -7,25 +7,24 @@ import com.instacart.formula.Transition
  * Note: this class is not a data class because equality is based on instance and not [key].
  */
 @PublishedApi
-internal class ListenerImpl<Input, State, Event>(internal var key: Any) : Listener<Event> {
+internal class ListenerImpl<Input, State, EventT>(internal var key: Any) : Listener<EventT> {
 
+    internal var manager: FormulaManagerImpl<Input, State, *>? = null
     internal var snapshotImpl: SnapshotImpl<Input, State>? = null
-    internal var transition: Transition<Input, State, Event>? = null
 
-    override fun invoke(event: Event) {
-        snapshotImpl?.let { snapshot ->
-            transition?.let { transition ->
-                val result = transition.toResult(snapshot, event)
-                snapshot.dispatch(result)
-                return
-            }
-        }
+    internal lateinit var transition: Transition<Input, State, EventT>
+
+    override fun invoke(event: EventT) {
         // TODO: log if null listener (it might be due to formula removal or due to callback removal)
+        val manager = manager ?: return
+
+        val deferredTransition = DeferredTransition(this, transition, event)
+        manager.onPendingTransition(deferredTransition)
     }
 
     fun disable() {
+        manager = null
         snapshotImpl = null
-        transition = null
     }
 }