+++ formula/src/test/java/com/instacart/formula/subjects/PendingActionFormulaTerminatedOnActionInit.kt
@@ -1,0 +1,67 @@
+package com.instacart.formula.subjects
+
+import com.instacart.formula.Action
+import com.instacart.formula.Evaluation
+import com.instacart.formula.Formula
+import com.instacart.formula.Snapshot
+import com.instacart.formula.StatelessFormula
+import com.instacart.formula.rxjava3.RxAction
+import com.instacart.formula.test.CountingInspector
+import com.instacart.formula.test.TestFormulaObserver
+import com.instacart.formula.test.TestableRuntime
+import io.reactivex.rxjava3.core.Observable
+
+class PendingActionFormulaTerminatedOnActionInit(runtime: TestableRuntime) {
+
+    private val inspector = CountingInspector()
+    private var observer: TestFormulaObserver<Unit, Int, ParentFormula>? = null
+    private val actionFormula = ActionFormula(terminateAction = {
+        observer?.dispose()
+    })
+
+    val test = runtime.test(ParentFormula(actionFormula), inspector).apply {
+        observer = this
+    }
+
+    fun assertActionsStarted(expected: Int) = apply {
+        inspector.assertActionsStarted(expected)
+    }
+
+    class ParentFormula(val actionFormula: ActionFormula): StatelessFormula<Unit, Int>() {
+        override fun Snapshot<Unit, Unit>.evaluate(): Evaluation<Int> {
+            val output = context.child(actionFormula)
+            return Evaluation(output)
+        }
+    }
+
+
+    class ActionFormula(val terminateAction: () -> Unit) : Formula<Unit, Int, Int>() {
+
+        override fun initialState(input: Unit): Int = 0
+
+        override fun Snapshot<Unit, Int>.evaluate(): Evaluation<Int> {
+            return Evaluation(
+                output = state,
+                actions = context.actions {
+                    RxAction.fromObservable {
+                        terminateAction()
+                        Observable.empty()
+                    }.onEvent { none() }
+
+                    // Pending actions
+                    Action.onInit().onEvent {
+                        transition(state + 1)
+                    }
+
+                    Action.onInit().onEvent {
+                        transition(state + 1)
+                    }
+
+                    Action.onInit().onEvent {
+                        transition(state + 1)
+                    }
+                }
+            )
+        }
+    }
+}